requirements.txt
#######################################################
### Database                                        ###
#######################################################
dj-database-url
djangorestframework

messagebird
googlemaps
